requirements.txt
python==3.12
langchain-groq==0.2.0
python-dotenv==1.0.1
chromadb==0.5.15
langchain-community==0.3.3
langchain-core==0.3.13
pandas==2.2.3S